pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "hatch-vcs"]

[project]
dependencies = [
  "bcrypt ~= 4.0.0",
  "fastapi ~= 0.89.0",
  "pydantic ~= 1.10.0",
  "pyjwt ~= 2.6.0",
  "python-multipart ~= 0.0.5",
  "uvicorn ~= 0.20.0",
]
dynamic = ["version"]
name = "api-auth"
readme = "README.md"
requires-python = "~=3.10.0"

[project.optional-dependencies]
dev = [
  "black ~= 22.12.0",
  "isort ~= 5.11.0",
  "httpx ~= 0.23.0",
  "mypy == 0.991",
  "pytest ~= 7.2.0",
  "pytest-cov ~= 4.0.0",
  "ruff ~= 0.0.224",
]

[tool.hatch.version]
source = "vcs"

[tool.isort]
profile = "black"

[tool.mypy]
strict = true
